Edit recurring hours and maybe extras

+++ app/src/main/java/com/handybook/handybook/booking/bookingedit/model/BookingEditHoursRequest.java
@@ -5,11 +5,11 @@ import com.google.gson.annotations.SerializedName;
 public final class BookingEditHoursRequest {
 
     @SerializedName("new_hrs")
-    private float mNewBaseHrs;
+    private double mNewBaseHrs;
     @SerializedName("apply_to_recurring")
     private boolean mApplyToRecurring;
 
-    public void setNewBaseHrs(final float newBaseHrs) {
+    public void setNewBaseHrs(final double newBaseHrs) {
         mNewBaseHrs = newBaseHrs;
     }
